% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotum leegnd.R
\name{add_legend}
\alias{add_legend}
\title{Beschriftung zu Graphik hinzufügen}
\usage{
add_legend(..., grid = TRUE, intervall = 0.5, lty = 3)
}
\arguments{
\item{...}{der Inhalt der Legende, meist legend=c(), dazu die Position der Legende ("right", "left", "top", "bottom" inkl. Kombinationen)}

\item{grid}{logisch: soll ein Gitter gezeichnet werden}

\item{intervall}{Abstand der Gitterlinien}

\item{lty}{Art der Gitterlinien (voreingestellt 3)}
}
\value{
Legend
}
\description{
Beschriftung zu Graphik hinzufügen
}
\examples{
add_legend("right", legend=c("Lotta", "Marie", "Olivia"), lty=c(1:3), col=c("steelblue", "indianred", "pink"), horiz=F, bty='n', cex=0.9) # hier die Beschriftung
}

