% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/row.has.na.R
\name{anzahlmissings}
\alias{anzahlmissings}
\title{Anzahl der Missings pro Fall}
\usage{
anzahlmissings(data)
}
\arguments{
\item{data}{Datensatz}
}
\value{
Anzahl der Missings pro Fall
}
\description{
Anzahl der Missings pro Fall
}
\examples{
demoframe <- data.frame(sapply(1:10,function(x) rnorm(30,100,10)))
demoframe <- as.data.frame(lapply(demoframe, function(x) "is.na<-"(x, sample(seq(x), floor(length(x) * runif(1, 0, .2))))))
demoframe
anzahlmissings(demoframe) # Anzahl Missings pro Fall
}

