% Generated by roxygen2 (4.0.0): do not edit by hand
\name{buildscale}
\alias{buildscale}
\title{<brief desc>}
\usage{
buildscale(faobject, ladungsunterschied = 0.1, datensatz = NULL,
  name = "skala", fuerntratt = FALSE)
}
\arguments{
\item{faobject}{saved fa object}

\item{ladungsunterschied}{should loadings be assigned to a factor by loading difference}

\item{datensatz}{data.frame containing the items}

\item{name}{Name the objects to be saved}

\item{fuerntratt}{should loading be assigned to factor according to Fuerntratt}
}
\value{
list of scales
}
\description{
<full description>
}
\examples{
\dontrun{

}
}
\seealso{
fa
}
\keyword{fa}

