% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colreorder.R
\name{colreorder}
\alias{colreorder}
\title{Reorder colums in a dataframe}
\usage{
colreorder(refcols, data)
}
\arguments{
\item{refcols}{columns to put at the left side or right side of a dataframe}

\item{data}{dataframe}
}
\value{
a dataframe with the reordered columns
}
\description{
to quickly reorder columns by name
}
\examples{
data(crime_data)
temp <- colreorder(c("crime","murder"), crime_data)
head(temp)
rm(temp)
}

