% Generated by roxygen2 (4.0.1): do not edit by hand
\name{cor_lm_d}
\alias{cor_lm_d}
\title{semipartial, partial und Korrelationen nullter Ordnung eines Regressionsmodell}
\usage{
cor_lm_d(fit)
}
\arguments{
\item{fit}{lm object}
}
\value{
dataframe of r, r.,r..
}
\description{
<full description>
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{lm},\link{cor}}
}
\keyword{linear}
\keyword{lm,}
\keyword{model,}
\keyword{regression}

