% Generated by roxygen2 (4.0.2): do not edit by hand
\name{corstarsmd_d}
\alias{corstarsmd_d}
\title{Korrelationstabelle mit Signifikanzsternen}
\usage{
corstarsmd_d(x, type = "markdown", stellen = 3, zehnproz = F, abk = T,
  diagonale = F, ...)
}
\arguments{
\item{x}{Datensatz (dataframe) Objekt mit den Variablen}

\item{type}{Ausgabevarianten ohne, html or markdown (übermittelt Tabelle an kable)}

\item{stellen}{Anzahl Stellen nach dem Komma}

\item{zehnproz}{(logisch) soll p<.10 mit einem Kreuz angezeigt werden}

\item{abk}{(logisch) sollen Spaltennamen gekürzt werden}

\item{diagonale}{(logisch) soll die Diagonale gefüllt werden}
}
\value{
text table
}
\description{
die Funktion errechnet eine Korrelationstabelle aus den übermittelten Variablen und fügt, wo angebracht, Signifikanzsterne hinzu
}
\examples{
data(pers_data)
corstarsmd_d(pers_data[c(7:10,12:15)],type="ohne", stellen=2, zehnproz=TRUE)

\dontrun{
# um das Ergebnis z.B. in Microsoft Word einzufügen, folgende Schritte ausführen:
for_word <- corstarsmd_d(pers_data,type = "ohne",stellen=2) # Ausgabe in Objekt sichern
library(xlsx) # Excel library laden
write.xlsx(for_word, "test.xlsx") # Ausgabe in ein Excel-Dokument speichern
# Das gespeicherte Dokument kann in Excel geöffnet werden,
# von dort Tabelle kopieren und in Word einfügen
}
}
\seealso{
\link{cor}
}
\keyword{correlation}

