% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{crime_data}
\alias{crime_data}
\title{Statewide Crime Data}
\format{A data frame with 51 rows and 9 variables}
\source{
A. Agresti and B. Finlay, STATISTICAL METHODS FOR THE SOCIAL SCIENCES, Prentice Hall, 3rd edition, 2008
}
\usage{
data(crime_data)
}
\description{
The data are from Statistical Abstract of the United States and most
variables were measured in 1993.
}
\details{
\itemize{
  \item \emph{sid} ID number
  \item \emph{state} state abbreviation
  \item \emph{crime} violent crime rate (per 100,000 people in population)
  \item \emph{murder} murder rate (per 100,000 people in population)
  \item \emph{pctmetro} percent in metropolitan areas
  \item \emph{pctwhite} percent white
  \item \emph{pcths} percent high school graduates
  \item \emph{poverty}  percent with income below the poverty level
  \item \emph{single} percent of families headed by a single parent
}
}
\keyword{datasets}

