% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinOmega_sq.R
\name{etasq}
\alias{etasq}
\title{Eta Quadrat und partielles Eta Quadrat einer Anova ausgeben}
\usage{
etasq(Anova3mod)
}
\arguments{
\item{Anova3mod}{}
}
\value{
η², peta2
}
\description{
Mit dieser Funktion kann die Effektstärke η²und pEta2 einer ANOVA berechnet werden
}
