% Generated by roxygen2 (4.0.0): do not edit by hand
\name{extractlavaan2table}
\alias{extractlavaan2table}
\title{given a lavaan multigroup object returns a dataframe to be used in publication}
\usage{
extractlavaan2table(inva = result)
}
\arguments{
\item{inva}{a lavaan object of an invariance test}
}
\value{
dataframe
}
\description{
<full description>
}
\examples{
\dontrun{
library(semTools)
inva <- measurementInvariance(model_reak,laurasdaten,group="Geschlecht",estimator="WLSMV")
extractlavaan2table(inva)
gsub(" 0.000", " <.001",print(xtable(extractlavaan2table(inva),digits=3), type="latex", sanitize.text.function=identity),useBytes=T)
}
}
\seealso{
lavaan
}
\keyword{invariance}
\keyword{lavaan,}

