% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/row.has.na.R
\name{fall.hat.missings}
\alias{fall.hat.missings}
\title{Missinganalyse pro Zeile}
\usage{
fall.hat.missings(data, logisch = FALSE)
}
\arguments{
\item{data}{Datensatz}

\item{logisch}{logisch: sollen ein logischer Vektor ausgegeben werden (standard=FALSE)}
}
\value{
Fallnummern bzw. Ja/Nein
}
\description{
Missinganalyse pro Zeile
}
\examples{
demoframe <- data.frame(sapply(1:10,function(x) rnorm(30,100,10)))
demoframe <- as.data.frame(lapply(demoframe, function(x) "is.na<-"(x, sample(seq(x), floor(length(x) * runif(1, 0, .2))))))
demoframe
fall.hat.missings(demoframe) # Fallnummern mit Missings
fall.hat.missings(demoframe,logisch = T) # Vektor T/F für Fälle mit und ohne Missings
anzahlmissings(demoframe)
}
