% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Fmax.R
\name{fmaxtest}
\alias{fmaxtest}
\title{F-Max Test für ANOVA}
\usage{
fmaxtest(erglm, deskriptive = TRUE)
}
\arguments{
\item{erglm}{lm Objekt}

\item{...}{}
}
\value{
Fmax, N-Ratio
}
\description{
\itemize{
\item Sind die Gruppen-Stichproben gleich groß bzw. ungleich groß bis zu einer Ratio von 4:1 (größte Stichprobe 4 mal größer als die kleinste Stichprobe), sollte der Fmax-Wert kleiner als 10 sein!
\item Bei ungleich großen Stichproben bis zu einer Ratio von 9:1 sollte der Fmax-Wert kleiner als 3 sein.
\item Wenn der Fmax-Test über den Grenzwerten liegt, dann erfolgt eine Reduktion des α-Niveaus:
\item Fmax > 10 bei Ratio 4:1  α < .025
\item Fmax > 3   bei Ratio 9:1  α < .025
}
}
\examples{
fmaxtest(erglm)
}
