% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/htmlschnell.R
\name{htmlschnell}
\alias{htmlschnell}
\title{Schnelle Darstellung von Data.frames als HTML}
\usage{
htmlschnell(dataframe)
}
\arguments{
\item{dataframe}{}
}
\value{
HTML
}
\description{
Mit dieser einfachen Funktion können Ergebnisse und Daten, die als data.frame vorliegen, 
als HTML dargestellt und so einfach in Word kopiert werden.
}
\examples{
data("crime_data")
describe(crime_data) \%>\% print(digits=2) \%>\% htmlschnell()
# einen Datensatz direkt
htmlschnell(crime_data[1:10,])
# eine Regression
lm_coef_spss_d(fit,stellen = 2) \%>\% htmlschnell()
}
