% Generated by roxygen2 (4.0.2): do not edit by hand
\docType{data}
\name{ist_data}
\alias{ist_data}
\title{Intelligence, School grades and life data}
\format{A data frame with 128 rows and 39 variables}
\usage{
data(ist)
}
\description{
The data comprises items from the IST-2000-R
its verbal, figural, numerical intelligence scores
school grades from math, physics, german, etc. and GPA
plus life data like age, sex, location
}
\details{
\itemize{
  \item vp \emph{VP ID}
  \item sex \emph{Geschlecht}
  \item age \emph{Alter}
  \item verbint \emph{verbale Intelligenz}
  \item numint \emph{num Intelligenz}
  \item figint \emph{fig Intelligenz}
  \item schlusint \emph{reasoning, schlussfolgerndes Denkes}
  \item se  an	gem	zr	re	rz	fa	wue	ma \emph{9 tests from the IST}
  }
}
\keyword{datasets}

