% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KontrastAbh.R
\name{kontrastAbh}
\alias{kontrastAbh}
\title{Kontrastanalyse f체r abh채ngige Stichproben}
\usage{
kontrastAbh(daten, kontrast)
}
\arguments{
\item{daten}{}

\item{kontrast}{}
}
\value{
data.frame
}
\description{
Kontrastanalyse f체r abh채ngige Stichproben
}

