% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KontrastvergleichAbh.R
\name{kontrastVergleichAbh}
\alias{kontrastVergleichAbh}
\title{Kontrastvergleich f체r abh채ngige Stichproben}
\usage{
kontrastVergleichAbh(daten, kontrast1, kontrast2)
}
\arguments{
\item{daten}{}

\item{kontrast}{}
}
\value{
data.frame
}
\description{
Kontrastvergleich f체r abh채ngige Stichproben
}
