% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/umittr-package.R
\docType{data}
\name{laurasdaten}
\alias{laurasdaten}
\title{18 items of a reactance questionnaire}
\format{A data frame with 524 rows and 19 variables}
\usage{
data(laurasdaten)
}
\description{
The data comprises 18 items measuring reactance
and sex (i.e. 'Geschlecht')
Items are scored from 1 to 4.
}
\keyword{datasets}

