% Generated by roxygen2 (4.0.1): do not edit by hand
\docType{package}
\name{lillionscage-package}
\alias{lillionscage}
\alias{lillionscage-package}
\title{Helper functions to ease the transition from SPSS to R}
\description{
Helper functions to ease the transition from SPSS to R
}
\author{
\email{macx@lillion.e4ward.com}
}
\seealso{
\code{\link{...}}
}
\keyword{package}

