% Generated by roxygen2 (4.0.2): do not edit by hand
\name{lm_coef_spss}
\alias{lm_coef_spss}
\title{linear regression coefficients output similar to SPSS}
\usage{
lm_coef_spss(fit)
}
\arguments{
\item{fit}{result of a lm}
}
\value{
dataframe with lm output similar to SPSS
}
\description{
function that returns a table of the regression coefficients similar to the presentation in SPSS
}
\examples{
library(car)
l <- lm(mpg ~ disp + hp + wt + drat, data=mtcars)
lm_coef_spss(l)

data(crime_data)
l <- lm(crime ~ pctwhite + pctmetro + poverty, data=crime_data)
lm_coef_spss(l)
rm(l)
\dontrun{
lapply(ll, lm_output_spss) # get outputs from several lm objects
}
}
\keyword{lm}
\keyword{regression}
\keyword{spss}

