% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aov_deskriptive.R
\name{lm_deskriptiv}
\alias{lm_deskriptiv}
\title{Deskriptive Statistik pro Gruppe einer Anova}
\usage{
lm_deskriptiv(lm)
}
\arguments{
\item{lm}{}
}
\value{
Deskriptive Statistik pro Gruppe
}
\description{
Deskriptive Statistik pro Gruppe einer Anova
}

