% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm_hierarch_model_summary.R
\name{lm_hier_model_sum}
\alias{lm_hier_model_sum}
\title{Hierarchical Regression model summary}
\usage{
lm_hier_model_sum(ll)
}
\arguments{
\item{ll}{a list containing regular lm objects}
}
\value{
dataframe of summaries
}
\description{
given a hierarchical regression with several blocks, gives an overview over changes in R^2 and significance
}
\examples{
library(car)
ll <- lm_hierarch(mpg ~ disp + hp + wt + drat, c(1,1,1,1), data=mtcars,summary=FALSE)
lm_hier_model_sum(ll)
rm(ll)
}
\seealso{
\code{\link{lm}}, \code{\link{lm_hierarch}}
}
\keyword{lm}
\keyword{regression}

