% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_hierarch}
\alias{lm_hierarch}
\title{Hierarchical regression with several blocks estimate}
\usage{
lm_hierarch(formula, blocks, data, summary = FALSE, ...)
}
\arguments{
\item{formula}{the complete formula including all blocks from the hierarchical regression}

\item{blocks}{numeric vector denoting the length of each block e.g. c(2, 3, 1) meaning 2 predictors then 3 predictors then 1 predictor}

\item{data}{the dataframe holding the variables for the estimation of the lm}

\item{summary}{should the results be summaries of the lm object}

\item{...}{additional parameters for \code{\link{lm}}}
}
\value{
list of the regression models
}
\description{
Hierarchichal regression with several blocks estimated in sequence
This is basically a wrapper function for \code{\link{lm}} that splits the formula
into user denoted blocks and computes its models in order to evaluate the incremental validity of each predictor set.
}
\examples{
\dontrun{
lm_hierarch(gewissen ~ sex + age + school, c(1),data=neoffi)
}
library(car)
lm_hierarch(mpg~disp+hp+wt+drat, c(1,1,1,1), data=mtcars) # 4 blocks with each time one additional predictor
lm_hierarch(mpg~disp+hp+wt+drat, c(2,2), data=mtcars) # 2 blocks with each time 2 additional predictors
}
\seealso{
\code{\link{lm}}
}
\keyword{hierarchical}
\keyword{lm,}
\keyword{regression}
\keyword{regression,}

