% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_hierarch_model_summary_d.R
\name{lm_hierarch_summary_d}
\alias{lm_hierarch_summary_d}
\title{Uebersicht der Ergebnisse einer Hierarchischen Regression}
\usage{
lm_hierarch_summary_d(formula, blocks, stellen = 2, ...)
}
\arguments{
\item{formula}{die Formel des vollständigen, d.h. alle Blöcke umfassenden Modells}

\item{blocks}{ein Vektor mit der Prädiktorenanzahl pro Block im Format c(Anzahl1,Anzahl2,…)}

\item{stellen}{auf wie viele Stellen soll der Output gerundet werden? Voreingestellt 2.}

\item{...}{Optionen, die an \code{\link{lm}} weitergegeben werden sollten, z.B. data=}
}
\value{
list mit den Ergebnissen
}
\description{
Diese Funktion führt eine hierarchische Regression mit mehreren Blöcken durch und bietet sowohl eine Modellübersicht mit den Veränderungen in R² und deren Signifikanz, als auch eine Übersicht der Koeffizienten pro Modell
}
\examples{
library(car)
lm_hierarch_summary_d(formula = mpg ~ disp + hp + wt + drat, 
blocks = c(1,1,1,1), # 4 Blöcke mit je einem zusätzlichen Prädiktor
data=mtcars)
}
\seealso{
\code{\link{lm}}, \code{\link{lm_hierarch_d}}
}
\keyword{lm}
\keyword{regression}
