% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_output_spss}
\alias{lm_output_spss}
\title{<brief desc>}
\usage{
lm_output_spss(fit)
}
\arguments{
\item{fit)}{}
}
\value{
dataframe with lm output similar to SPSS
}
\description{
<full description>
}
\examples{
\dontrun{

}
}
\keyword{lm,}
\keyword{regression,}
\keyword{spss}

