% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lm_output_spss_d}
\alias{lm_output_spss_d}
\title{Lineare Regression ähnlich der aus SPSS}
\usage{
lm_output_spss_d(fit)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}
}
\value{
dataframe des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
<full description>
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{lm}}
}
\keyword{lm,}
\keyword{regression,}
\keyword{spss}

