% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_output_spss_wrapper_d}
\alias{lm_output_spss_wrapper_d}
\title{Ergebnis einer linearen Regression ähnlich SPSS}
\usage{
lm_output_spss_wrapper_d(fit, Rsquare = TRUE, coefficients = TRUE,
  collinearity = TRUE, plot = TRUE)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}

\item{plot}{should histogram of residuals, qq plot and predicted/residual value plot be shown}
}
\value{
dataframe des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
<full description>
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{lm}}
}
\keyword{lm,}
\keyword{regression,}
\keyword{spss}

