% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_pred_r_squared}
\alias{lm_pred_r_squared}
\title{Predictive R-squared}
\usage{
lm_pred_r_squared(l)
}
\arguments{
\item{linear.model}{A linear regression model (class 'lm'). Required.}
}
\description{
returns the predictive r-squared using
             the PRESS statistic.
}
\author{
Thomas Hopper
}

