% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/mgsub.R
\name{mgsub}
\alias{mgsub}
\title{Buchstabenreihen ersetzen in Strings}
\usage{
mgsub(pattern, replacement, x, ...)
}
\arguments{
\item{pattern}{das Muster nach Art von c("ä","ü","ö")}

\item{replacement}{Ersatzmuster nach Art von c("ae","ue","oe")}

\item{x}{Stringvektor, in dem die Ersetzung stattfinden soll}

\item{...}{weitere Optionen für gsub}
}
\value{
string vektor
}
\description{
<full description>
}
\examples{
\dontrun{
mgsub(c("ä","ü","ö","Ä","Ü","Ö","ß"), c("ae","ue","oe","Ae","Ue","Oe","ss"), beschriftung)
}
}
\seealso{
\code{\link{gsub}}
}
\keyword{gsub}

