% Generated by roxygen2 (4.0.2): do not edit by hand
\name{moderation.slopes}
\alias{moderation.slopes}
\title{calculates slopes and intercepts for a moderation and plots a simple graph}
\usage{
moderation.slopes(mod, zsd = 1, mod_name = "Moderator", ...)
}
\arguments{
\item{mod}{the result of a moderation
done like this: lm(y~x*z), so that you have one predictor x
and one moderator z}

\item{zsd}{the standard deviation of z to calculate the slopes}

\item{...}{Parameter for plot (xlab, ylab, main etc.)}
}
\value{
moderation
}
\description{
calculates the slopes and intercepts for a simple moderation analysis from a lm
}
\examples{
data(crime_data)
attach(crime_data)
xm <- poverty-mean(poverty, na.rm=T)
zm <- pcths-mean(pcths, na.rm=T)
zz <- xm*zm
mod <- lm(crime~poverty+pcths+zz, data=crime_data)
moderation.slopes(mod, xlab="crime rate")
detach(crime_data)
}
\seealso{
\code{\link{lm}}
}
\keyword{lm}

