% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multibar_multihist.R
\name{multibar}
\alias{multibar}
\title{Multiple separate barplots}
\usage{
multibar(x, variables, labels, xlab = "Antworten", save = TRUE, ...)
}
\arguments{
\item{x}{data.frame}

\item{variables}{vector of variables to be used}

\item{labels}{labels for each seperate histogram}

\item{xlab}{general subscript for each histogram}

\item{save}{should pdf be saved to disk}

\item{...}{additional graphic parameters for barplot (e.g., ylab col)}
}
\value{
Generates for each variable in a dataframe a barplot, optionally saved to the wd as a pdf
}
\description{
given a dataframe, produces a barplot for each variable or a subset of variables, which optionally can be saved to disk as a pdf
it's actually just a wrapper for \code{\link{barplot}}
}
\examples{
data(cars)
multibar(cars,save=FALSE)
}
\seealso{
\code{\link{multihist}} \code{\link{barplot}}
}
\keyword{plot}
\keyword{umit}

