% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multibar_d}
\alias{multibar_d}
\title{Mehrere getrennte barplots}
\usage{
multibar_d(x, variables, labels, xlab = "Antworten", save = TRUE, ...)
}
\arguments{
\item{x}{data.frame}

\item{variables}{Vektor der Variablen, die dargestellt werden sollen}

\item{labels}{String-Vektor der Labels, die verwendet werden sollen}

\item{xlab}{Unterschrift unter dem Plot}

\item{save}{Soll ein PDF gespeichert werden}

\item{...}{zusätzliche Parameter von hist (z.B. col)}
}
\value{
Barplot
}
\description{
given a dataframe, produces a barplot for each variable or a subset of variables, which optionally can be saved to disk as a pdf
}
\examples{
data(cars)
multibar(cars,save=F)
}
\seealso{
mulithist
}
\keyword{umit}

