% Generated by roxygen2 (4.0.0): do not edit by hand
\name{multihist}
\alias{multihist}
\title{<brief desc>}
\usage{
multihist(x, variables, labels = TRUE, xlab = "Verteilung", save = TRUE,
  ...)
}
\arguments{
\item{x}{}

\item{variables}{}

\item{labels}{}

\item{xlab}{}
}
\value{
none
}
\description{
<full description>
}
\examples{
\dontrun{
multihist(dd)
multibar(dd, 80:82, save = T)
multibar(x2,c(1:3))
multibar(neo_n,5,ylab="häu")
}
}
\seealso{
lillionscage::multibar
}
\keyword{umit}

