% Generated by roxygen2 (4.0.2): do not edit by hand
\name{multihist_d}
\alias{multihist_d}
\title{Mehrere getrennte Histogramme}
\usage{
multihist_d(x, variables, labels = TRUE, xlab = "Verteilung", save = TRUE,
  ...)
}
\arguments{
\item{x}{data.frame}

\item{variables}{Vektor der Variablen, die dargestellt werden sollen}

\item{labels}{String-Vektor der Labels, die verwendet werden sollen}

\item{xlab}{Unterschrift unter dem Plot}

\item{save}{Soll ein PDF gespeichert werden}

\item{...}{zusätzliche Parameter von hist (z.B. col)}
}
\value{
Histogramme
}
\description{
produces a histogram for each variable in a dataframe or its subset, which optionally can be saved to disk as a pdf
}
\examples{
\dontrun{
multihist(dd)
multibar(dd, 80:82, save = T)
multibar(x2,c(1:3))
multibar(neo_n,5,ylab="häu")
}
}
\seealso{
\code{\link{multibar_d}}
}
\keyword{Graphik}
\keyword{umit}

