% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corrplotumit.R
\name{myPlotumit}
\alias{myPlotumit}
\title{Korrelationsplot für einfache Regression}
\usage{
myPlotumit(beta, data)
}
\arguments{
\item{beta}{data}
}
\value{
plot
}
\description{
dient nur Lernzwecken
}
\examples{
library(dplyr)
library(psych)
dfcov <- data_frame(punkte_intel=c(10,30,10,50,30,15),
punkte_klaus=c(10,19,30,50,60,20),dx=punkte_intel-mean(punkte_intel),
dy=punkte_klaus-mean(punkte_klaus),covZ=dx*dy,cov=covZ/(length(punkte_intel)-1))
library(manipulate)
manipulate(myPlotumit(beta,dfcov), beta = slider(-1, 1, step = 0.01))
}
\seealso{
lm
}
\keyword{lm}

