% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neue Funktionen.R
\name{myscatter}
\alias{myscatter}
\title{schneller Scatterplot}
\usage{
myscatter(df, xvar, yvar)
}
\arguments{
\item{df}{}

\item{xvar}{}

\item{yvar}{}
}
\value{

}
\description{
schneller Scatterplot
}
\examples{
data(crime_data)
myscatter(crime_data,crime, pctmetro)
}
