% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/neue Funktionen.R
\name{mysummarise}
\alias{mysummarise}
\title{schnelle deskriptive Statistik mit se und CI}
\usage{
mysummarise(df, var)
}
\arguments{
\item{df}{}

\item{var}{}
}
\value{

}
\description{
schnelle deskriptive Statistik mit se und CI
}
\examples{
data(crime_data)
mysummarise(crime_data,crime)
}
