% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{odds2p}
\alias{odds2p}
\title{Odds zu Wahrscheinlichkeiten umwandeln}
\usage{
odds2p(odds)
}
\arguments{
\item{odds}{einzeln oder als Vektor oder als glm Objekt}
}
\value{
p Wert
}
\description{
Odds zu Wahrscheinlichkeiten umwandeln
}
\examples{
odds2p(c(1,2,.5))
}
