% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meinOmega_sq.R
\name{omegasq}
\alias{omegasq}
\title{Omega Quadrat einer Anova ausgeben}
\usage{
omegasq(Anova3mod)
}
\arguments{
\item{Anova3mod}{}
}
\value{
omega2
}
\description{
Mit dieser Funktion kann die Effektstärke Omega2 einer ANOVA berechnet werden
}
