% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper.R
\name{p2odds}
\alias{p2odds}
\title{Wahrscheinlichkeiten (p) in Odds umwandeln}
\usage{
p2odds(p)
}
\arguments{
\item{p}{}
}
\value{
Odds
}
\description{
Wahrscheinlichkeiten (p) in Odds umwandeln
}
\examples{
p2odds(c(.10,.50,.90))
}

