% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/umittr-package.R
\docType{data}
\name{pers_data}
\alias{pers_data}
\title{Personality data with life satisfaction}
\format{A data frame with 345 rows and 16 variables}
\usage{
data(pers_data)
}
\description{
The data comprises items and sum-scores for Extraversion
Neuroticism and Life Satisfaction
Items are scored from 1 to 5.
}
\details{
\itemize{
  \item lz1 \emph{In den meisten Bereichen entspricht mein Leben meinen Idealvorstellungen}
  \item lz2 \emph{Meine Lebensbedingungen sind ausgezeichnet}
  \item lz3 \emph{Ich bin mit meinem Leben zufrieden}
  \item lz4 \emph{Bisher habe ich die wesentlichen Dinge erreicht, die ich mir für mein Leben wünsche.}
  \item lz5 \emph{Wenn ich mein Leben noch einmal leben könnte, würde ich kaum etwas ändern.}
  \item lezu \strong{Skala }\emph{Lebenszufriedenheit}
  \item e1 \emph{kontaktfreudig}
  \item e2 \emph{gesellig}
  \item e3 \emph{lebhaft}
  \item e4 \emph{temperamentvoll}
  \item extra \strong{Skala }\emph{Extraversion}
  \item n1 \emph{verletzbar}
  \item n2 \emph{empfindlich}
  \item n3 \emph{launenhaft}
  \item n4 \emph{selbstzweiflerisch}
  \item neuro \strong{Skala }\emph{Neurotizismus}
  }
}
\keyword{datasets}

