% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotum legend.R
\name{plotum}
\alias{plotum}
\title{Plot Funktion mit einigen Voreinstellungen zur leichteren Handhabung}
\usage{
plotum(
  ...,
  position = "right",
  bty = "n",
  las = 1,
  cex.axis = 0.8,
  Schrift = "Garamond"
)
}
\arguments{
\item{...}{}

\item{position}{}

\item{bty}{}

\item{las}{}

\item{cex.axis}{}

\item{Schrift}{}
}
\value{

}
\description{
Plot Funktion mit einigen Voreinstellungen zur leichteren Handhabung
}
\examples{
plotum(23:27,4:8,Schrift="Gill Sans")
add_legend("right", legend=c('Text hier', 'Text2 hier', '...'), lty=c(1:3), col=c('pink','indianred', 'steelblue'), horiz=FALSE, bty='n', cex=0.9)
}
