% Generated by roxygen2 (4.0.2): do not edit by hand
\name{r2_stelzl}
\alias{r2_stelzl}
\title{compute corrected r2 and omega2 for a regression}
\usage{
r2_stelzl(lmm)
}
\arguments{
\item{lmm}{}
}
\value{
r2, omega2
}
\description{
compute corrected r2 and omega2 for a regression
as described by Stelzl (2005)
}
\examples{
\dontrun{

}
}
\seealso{
lm
}
\keyword{lm}

