% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Radar_Plot_Eni.R
\name{RadialPlot}
\alias{RadialPlot}
\title{Radarplot}
\usage{
RadialPlot(plot.data, axis.labels = colnames(plot.data)[-1], grid.min = 0,
  grid.mid = 4.5, grid.max = 10, centre.y = grid.min - ((1/9) * (grid.max
  - grid.min)), polygon = FALSE, plot.extent.x.sf = 1.2,
  plot.extent.y.sf = 1.2, x.centre.range = 0.02 * (grid.max - centre.y),
  label.centre.y = FALSE, grid.line.width = 0.5,
  gridline.min.linetype = "longdash", gridline.mid.linetype = "longdash",
  gridline.max.linetype = "longdash", gridline.min.colour = "grey",
  gridline.mid.colour = "blue", gridline.max.colour = "grey",
  gridline.mid.alpha = 1, grid.label.size = 4,
  gridline.label.offset = -0.02 * (grid.max - centre.y),
  label.gridline.min = FALSE, axis.label.offset = 1.15,
  axis.label.size = 3, axis.line.colour = "snow2", group.line.width = 1,
  group.point.size = 4, background.circle.colour = "yellow",
  background.circle.transparency = 0.2, plot.legend = if (nrow(plot.data) >
  1) TRUE else FALSE, legend.title = "Übersicht",
  legend.text.size = grid.label.size)
}
\arguments{
\item{plot.data}{Daten}

\item{axis.labels}{}

\item{grid.min}{Mindestwert der Daten bzw. Darstellung}

\item{grid.mid}{}

\item{grid.max}{Maximalwert der Daten bzw. Darstellung}

\item{centre.y}{meist auf 0 zentriert}

\item{polygon}{}

\item{plot.extent.x.sf}{}

\item{plot.extent.y.sf}{}

\item{x.centre.range}{}

\item{label.centre.y}{}

\item{grid.line.width}{}

\item{gridline.min.linetype}{}

\item{gridline.mid.linetype}{}

\item{gridline.max.linetype}{}

\item{gridline.min.colour}{}

\item{gridline.mid.colour}{}

\item{gridline.max.colour}{}

\item{gridline.mid.alpha}{}

\item{grid.label.size}{}

\item{gridline.label.offset}{}

\item{label.gridline.min}{}

\item{axis.label.offset}{}

\item{axis.label.size}{}

\item{axis.line.colour}{}

\item{group.line.width}{}

\item{group.point.size}{}

\item{background.circle.colour}{}

\item{background.circle.transparency}{}

\item{plot.legend}{}

\item{legend.title}{}

\item{legend.text.size}{}
}
\value{
Radar Graphik
}
\description{
Radarplot
}
\examples{
tb <- data.frame(group = c("Januar", "November"), Aktivität = c(3,
2.4), Interaktion = c(4L, 2L), Ankommen = c(4.2, 1.5), Gehen = c(3L, 1L),
Spiele = c(3.5, 3), Begrüßungslied = c(3.6, 2.8),
Stimmung = c(4L, 3L), Essverhalten = c(3, 1.8), Konfliktsituation = c(3.8,2.2))
RadialPlot(tb[1:6],centre.y = 0,label.gridline.min = F,group.line.width = .5,group.point.size = 2,background.circle.colour = "white",grid.label.size = 0,polygon = FALSE,plot.extent.x.sf = 1.3,legend.title = "Kind A")+scale_alpha(guide='none')
# mit Füllung:
RadialPlot(tb,grid.min = 0,grid.max = 4.5,centre.y = 0,label.gridline.min = F,group.line.width = .5,group.point.size = 2,background.circle.colour = "white",grid.label.size = 0,polygon = TRUE,plot.extent.x.sf = 1.3,axis.line.colour = "snow2")+
ggtitle("Entwicklung Kind A") +
  scale_fill_discrete("Kind A") +
    scale_colour_discrete("Kind A")+scale_alpha(guide='none')
}

