% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Normierung.R
\name{RohwertezuNormwerten}
\alias{RohwertezuNormwerten}
\title{Rohwerten Normwerte zuordnen}
\usage{
RohwertezuNormwerten(Rohwerte, Normwerte)
}
\arguments{
\item{Rohwerte}{}

\item{Normwerte}{}
}
\value{
Dataframe
}
\description{
Rohwerten Normwerte zuordnen
}
\examples{
x <- data.frame(x=rnorm(10000)) # Datensatz erzeugen
x$T <- Tscorestransformieren(x$x) # in T-Werte umwandeln
RohwertezuNormwerten(Rohwerte = x["x"],Normwerte = x["T"])
x$Stanine <- staninentransformieren(x$x) # in Stanine umwandeln
y <- x[RohwertezuNormwerten(Rohwerte = x$x,Normwerte = x["T"])$rows,]
y
}
\references{
Tent, L., Stelzl, I. (1993), Pädagogisch-psychologische Diagnostik: Theoretische und methodische Grundlagen. Göttingen: Hogrefe.
}

