% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/row.has.na.R
\name{row.has.na}
\alias{row.has.na}
\title{Analyse cases for missings}
\usage{
row.has.na(data, which = TRUE, percase = FALSE)
}
\arguments{
\item{data}{dataframe}

\item{which}{logical: should only case/row numbers be displayed}

\item{percase}{logical: show missings per case/row}
}
\value{
case numbers or LOGICAL
}
\description{
Analyse cases for missings
}
\examples{
demoframe <- data.frame(sapply(1:15,function(x) rnorm(50,100,10)))
demoframe <- as.data.frame(lapply(demoframe, function(x) "is.na<-"(x, sample(seq(x), floor(length(x) * runif(1, 0, .2))))))
row.has.na(demoframe)
}

