% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{erstes Setup von R f?r das RProfile}
\usage{
setup()
}
\value{
schreibt eine User-Datei auf die Festplatte.
Falls die Datei '.Rprofile' schon besteht, wird ein Backup
der alten Daten erzeugt.
}
\description{
erstes Setup von R f?r das RProfile
}
\examples{
setup()
# wenn es funktioniert hat:
o() # öffnet das aktuelle Verzeichnis
}

