% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/simplebootstrap.R
\name{simpleboot}
\alias{simpleboot}
\title{simple univariate bootstrap}
\usage{
simpleboot(x, stat, reps = 1000)
}
\arguments{
\item{x}{vector of data}

\item{stat}{which stat should be bootstrapped, e.g. 'mean', 'sd'}

\item{reps}{number of repeats}
}
\value{
bootstrap
}
\description{
simple bootstrap for simple univariate statistics
}
\examples{
mydata<-rchisq(25,df=3)
simpleboot(mydata,"mean",reps=10000)
}
\seealso{
boot
}
\keyword{bootstrap}

