% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplebootstrap_d.R
\name{simpleboot_d}
\alias{simpleboot_d}
\title{Einfacher univariater Bootstrap}
\usage{
simpleboot_d(x, stat, reps = 1000)
}
\arguments{
\item{x}{Vektor mit den Daten}

\item{stat}{Art der Statistik, z.B. 'mean', 'sd'}

\item{reps}{Anzahl der Wiederholung}
}
\value{
bootstrap
}
\description{
einfacher bootstrap für  univariate Statistik
}
\examples{
mydata<-rchisq(25,df=3)
simpleboot_d(mydata, "mean", reps=10000)
simpleboot_d(mydata, sd, reps=5000)
}
\seealso{
boot
}
\keyword{bootstrap}

