% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper.R
\name{spss}
\alias{spss}
\title{spss einflesen
wrapper für die foreign Funtion}
\usage{
spss(file)
}
\arguments{
\item{file}{Pfad zur Datei}
}
\value{
data.frame
}
\description{
spss einflesen
wrapper für die foreign Funtion
}

