% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Normierung.R
\name{staninenormierung}
\alias{staninenormierung}
\title{Werten Stanine Normierung zuweisen}
\usage{
staninenormierung(x)
}
\arguments{
\item{x}{}
}
\value{
Dataframe
}
\description{
Werten Stanine Normierung zuweisen
}
\examples{
x <- rnorm(10000) # Datensatz erzeugen
y <- c(1.2,-1,.4,3) # für diese Werte
staninenormierung(x)
table(staninenormierung(x))
hist(staninenormierung(x), breaks=1:9,xlim=c(0,10), axis=F); axis(side = 1,labels = 1:9,tick = 1:9,at = 1:9)
}

