% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Normierung.R
\name{staninentransformieren}
\alias{staninentransformieren}
\title{Werten Stanine Normierung zuweisen}
\usage{
staninentransformieren(x)
}
\arguments{
\item{x}{}
}
\value{
Dataframe
}
\description{
Werten Stanine Normierung zuweisen
}
\examples{
x <- rnorm(10000) # Datensatz erzeugen
y <- c(1.2,-1,.4,3) # für diese Werte 
staninentransformieren(x)
table(staninentransformieren(x))
hist(staninentransformieren(x),breaks=0:9,xlim=c(0,9))
hist(exp(x)) # schief verteilte Daten
x <- staninentransformieren(exp(x)) 
table(staninentransformieren(x))
}

