% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/SummarySE.R
\name{summarySE}
\alias{summarySE}
\title{Standardfehler und Konfidenzintervall des Mittelwerts, auch gruppenbezogen}
\usage{
summarySE(
  data = NULL,
  measurevar,
  groupvars = NULL,
  na.rm = FALSE,
  conf.interval = 0.95,
  .drop = TRUE
)
}
\arguments{
\item{data}{}

\item{measurevar}{}

\item{groupvars}{}

\item{na.rm}{}

\item{conf.interval}{}

\item{.drop}{}
}
\value{

}
\description{
Standardfehler und Konfidenzintervall des Mittelwerts, auch gruppenbezogen
}
\examples{
summarySE(crime_data,"pcths")
summarySE(mtcars,"mpg",groupvars="cyl", conf.interval = 80)
}
