% Generated by roxygen2 (4.0.0): do not edit by hand
\name{tablemissing}
\alias{tablemissing}
\title{Plot different patterns of missing values of multiple items}
\usage{
tablemissing(x, sortby = "both", main = "Missing Value Plot", hue = 0.6,
  value = 0.9, cut = 0)
}
\arguments{
\item{x}{dataframe with multiple items}

\item{sortby}{should the plot be sorted by "row", i.e. pattern, "column", i.e. item, or "both"}

\item{main}{title for the plot}

\item{hue}{hue, i.e. base color, in the hsv color model}

\item{value}{value, i.e. luminance, in the hsv color model}

\item{cut}{minimum number of items a missing pattern has to contain to be included in the plot}
}
\value{
plot
}
\description{
Given a number of items, plot their individual number of missings and the patterns of missings across the items
}
\examples{
demoframe <- data.frame(sapply(1:15,function(x) rnorm(50,100,10)))
demoframe <- as.data.frame(lapply(demoframe, function(x) "is.na<-"(x, sample(seq(x), floor(length(x) * runif(1, 0, .2))))))
tablemissing(demoframe,cut=1)
}
\seealso{
table
}
\keyword{missings}

