% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tablemissing-mod-color_mod_d.R
\name{tablemissing_d}
\alias{tablemissing_d}
\title{Graphische Darstellung unterschiedlicher missing-value Muster ueber mehrere Items}
\usage{
tablemissing_d(x, sortby = "beides", main = "Darstellung der Missings",
  hue = 0.6, value = 0.9, cut = 0)
}
\arguments{
\item{x}{mehrere Items als Datensatz (dataframe)}

\item{sortby}{wonach soll die Darstellung sortiert werden: nach Muster -> "Muster", nach Item -> "Spalte",oder nach beidem "beides" (voreingestellt)}

\item{main}{Überschrift der Darstellung}

\item{hue}{Farbe (H-ue), d.h. Grundfarbe, im hsv Farbmodel}

\item{value}{Wert (V-alue), d.h. Helligkeit, im hsv Farbmodel}

\item{cut}{Minimale Häufigkeit eines Missing-Musters, damit es in der Graphik mitaufgenommen wird}
}
\value{
plot
}
\description{
Für eine Anzahl von Items werden die jeweilige Anzahl der Missing per Item
 als auch die unterschiedlichen Missing-Muster über alle Items graphisch dargestellt
}
\examples{
demoframe <- data.frame(sapply(1:15,function(x) rnorm(50,100,10)))
demoframe <- as.data.frame(lapply(demoframe, function(x) "is.na<-"(x, sample(seq(x), floor(length(x) * runif(1, 0, .2))))))
tablemissing_d(demoframe,cut=0) # Muster, die nur 1x vorkommen, werden unterdrückt
tablemissing_d(demoframe,cut=0,sortby="Muster")

}
\seealso{
table
}
\keyword{missings}
