% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_isb.R
\name{theme_isb}
\alias{theme_isb}
\alias{theme_isb_grey}
\alias{theme_isb_ohneAxis}
\alias{theme_min}
\title{Helles Thema für ggplot2}
\usage{
theme_isb()

theme_isb_grey()

theme_isb_ohneAxis()

theme_min(
  size = 10,
  font = NA,
  face = "plain",
  backgroundColor = "white",
  panelColor = "grey",
  axisColor = "#999999",
  gridColor = "grey",
  textColor = "black"
)
}
\arguments{
\item{size}{}

\item{font}{}

\item{face}{}

\item{backgroundColor}{}

\item{panelColor}{}

\item{axisColor}{}

\item{gridColor}{}

\item{textColor}{}
}
\value{







}
\description{
Helles Thema für ggplot2

Themen für ggplot2

Thema ohne Achsen

Title
}
\examples{
require(ggplot2)
p1 <- ggplot(tibble(x=rnorm(100),y=rnorm(100)^3),aes(x,y))+geom_point()+stat_smooth()+theme_isb()
p2 <- ggplot(tibble(x=rnorm(100),y=rnorm(100)^3),aes(x,y))+geom_point()+stat_smooth()+theme_isb_grey()
p3 <- ggplot(tibble(x=rnorm(100),y=rnorm(100)^3),aes(x,y))+geom_point()+stat_smooth()+theme_isb_ohneAxis()
cowplot::plot_grid(p1, p2, p3, labels = c("theme_isb","theme_isb_grey","theme_isb_ohneAxis"),hjust=-.1,scale=.85,label_size=12)
}
