% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/theme_isb.R
\name{theme_isb_oPfeil}
\alias{theme_isb_oPfeil}
\alias{theme_isb_oPfeil_quer}
\title{ggplot2 theme ohne Pfeile an den Achsen}
\usage{
theme_isb_oPfeil(Winkel = 90, hVer = 1)

theme_isb_oPfeil_quer(Winkel = 0, hVer = 0.5)
}
\arguments{
\item{Winkel}{}

\item{hVer}{}
}
\value{



}
\description{
ggplot2 theme ohne Pfeile an den Achsen

Title
}
\examples{
require(ggplot2)
p4 <- ggplot(tibble(x=rnorm(100),y=rnorm(100)^3),aes(x,y))+geom_point()+stat_smooth()+theme_isb_oPfeil()
p5 <- ggplot(tibble(x=rnorm(100),y=rnorm(100)^3),aes(x,y))+geom_point()+stat_smooth()+theme_isb_oPfeil_quer()
cowplot::plot_grid(p4, p5, labels = c("theme_isb_oPfeil","theme_isb_oPfeil_quer"))
ggplot(tibble(x=rnorm(100),y=rnorm(100)^3),aes(x,y))+geom_point()+stat_smooth()+stat_smooth(method="lm",color="red",fill="lightblue")+theme_isb_oPfeil_quer()
}
