% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/wrapper.R
\name{tprozent}
\alias{tprozent}
\title{T/SW Werte in Prozenz
wandelt einzelne oder mehrere T- bzw. S(W)-Werte in Prozent um}
\usage{
tprozent(twert, SW = FALSE)
}
\arguments{
\item{twert}{}

\item{SW}{logisch wenn TRUE, dann Standardwerte (MW 100, SD 10)}
}
\value{
Prozentwerte
}
\description{
T/SW Werte in Prozenz
wandelt einzelne oder mehrere T- bzw. S(W)-Werte in Prozent um
}
\examples{
tprozent(20:80)
}

