% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Normierung.R
\name{Tscores}
\alias{Tscores}
\title{T Werte eines Vektors berechen}
\usage{
Tscores(x)
}
\arguments{
\item{x}{Vektor}
}
\value{
Dataframe
}
\description{
T Werte eines Vektors berechen
}
\examples{
x <- data.frame(x=rnorm(10000)) # Datensatz erzeugen
Tscores(x$x) # Normwerte erzeugen
}

