% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/colreorder_d.R
\name{varumstellen}
\alias{varumstellen}
\title{Variablen-Reihenfolge ändern}
\usage{
varumstellen(variablen, data)
}
\arguments{
\item{variablen}{columns to put at the left side or right side of a dataframe}

\item{data}{Datensatz}
}
\value{
a dataframe with the reordered columns
}
\description{
für ein simples Umstellen der Variablenreihenfolge im Datensatz
}
\examples{
data(crime_data)
temp <- varumstellen(c("crime","murder"), crime_data)
head(temp)
rm(temp)
}

