% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/render_word.R
\name{wordrender}
\alias{wordrender}
\title{Rendern als Word-Dokument}
\usage{
wordrender(file, ...)
}
\arguments{
\item{file}{das R Skript, das gerendert werden soll}
}
\value{
Word Doc
}
\description{
Diese Funktion wandelt ein R-Skript in eine Word-Datei um. Dafür ist das Paket "rmarkdown" notwendig!
}
\examples{
KI(wert = 66,standardabweichung = 10,reliabilitaet = .98,sicherheit = .99,seitigkeit = 1,mittelwert = 50,typ = "r") # inklusive Regressionshypothese
KI(wert = 66,standardabweichung = 10,reliabilitaet = .96,sicherheit = .8,seitigkeit = 1,mittelwert = 50,typ = "r") # inklusive Regressionshypothese
}
