% Generated by roxygen2 (4.0.2): do not edit by hand
\name{cor_lm}
\alias{cor_lm}
\title{semipartial, partial and zero-order correlations for a LM}
\usage{
cor_lm(fit)
}
\arguments{
\item{fit}{lm object}
}
\value{
dataframe of r, r.,r..
}
\description{
compute semipartial, partial and zero-order correlations by way of an lm object
}
\examples{
library(car)
fit <- lm(mpg~disp+hp+wt+drat, data=mtcars)
cor_lm(fit)
}
\seealso{
lm()
}
\keyword{linear}
\keyword{lm,}
\keyword{model,}
\keyword{regression}

