% Generated by roxygen2 (4.0.0): do not edit by hand
\name{cor_lm}
\alias{cor_lm}
\title{semipartial, partial and zero-order correlations for a LM}
\usage{
cor_lm(fit)
}
\arguments{
\item{fit}{lm object}
}
\value{
dataframe of r, r.,r..
}
\description{
<full description>
}
\examples{
\dontrun{

}
}
\seealso{
lm()
}
\keyword{linear}
\keyword{lm,}
\keyword{model,}
\keyword{regression}

