% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/semi_partial_corr_lm_d.R
\name{cor_lm_d}
\alias{cor_lm_d}
\title{semipartial, partial und Korrelationen nullter Ordnung eines Regressionsmodell}
\usage{
cor_lm_d(fit)
}
\arguments{
\item{fit}{ein Objekt einer \code{\link{lm}} Berechnung}
}
\value{
Datensatz (dataframe) von r, r.,r..
}
\description{
berechnet semipartial, partial und Korrelationen nullter Ordnung der Prädiktoren einer linearen Regression mit dem Kriterium
}
\examples{
library(car)
fit <- lm(mpg ~ disp + hp + wt + drat, data=mtcars)
cor_lm_d(fit)
}
\seealso{
\code{\link{lm},\link{cor}}
}
\keyword{"linear}
\keyword{korrelation}
\keyword{lm}
\keyword{model"}
\keyword{regression}

