% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/corstarsmd.R
\name{corstarsmd}
\alias{corstarsmd}
\title{Correlation table with significance stars}
\usage{
corstarsmd(x, type = "markdown", digits = 3, tenperc = FALSE,
  abbrev = TRUE, diagonal = FALSE, lower = TRUE, ...)
}
\arguments{
\item{x}{dataframe object}

\item{type}{options are none, html or markdown (hands table to kable)}

\item{digits}{how many digits should be displayed}

\item{tenperc}{(logical) should p<.10 also be denoted by a bubble (°)}

\item{abbrev}{(logical) should columnames be shortened}

\item{diagonal}{(logical) should the diagonal be filled}

\item{lower}{(logical) should the lower triangle be used}

\item{...}{additional parameters for kable}
}
\value{
text table
}
\description{
provides a table of a correlation matrix including denoted significance
}
\examples{
data(pers_data)
corstarsmd(pers_data[c(7:10,12:15)],type="none", digits=2, tenperc=TRUE)
\dontrun{
# in order to transfer the output into e.g. Microsoft Word, use the following workflow:
for_word <- corstarsmd(pers_data,type = "none",digits=2)
library(xlsx)
write.xlsx(for_word, "test.xlsx")
# the saved file can be opened in Excel, the table copied and pasted into Word
}
}
\seealso{
\link{cor}
}
\keyword{correlation}
\keyword{semi}

