% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/corstarsmd_d.R
\name{corstarsmd_d}
\alias{corstarsmd_d}
\title{Korrelationstabelle mit Signifikanzsternen}
\usage{
corstarsmd_d(
  x,
  type = "markdown",
  stellen = 3,
  zehnproz = FALSE,
  abk = TRUE,
  diagonale = FALSE,
  untere = TRUE,
  ...
)
}
\arguments{
\item{x}{Datensatz (dataframe) Objekt mit den Variablen}

\item{type}{Ausgabevarianten ohne, html or markdown (übermittelt Tabelle an kable)}

\item{stellen}{Anzahl Stellen nach dem Komma}

\item{zehnproz}{(logisch) soll p<.10 mit einem Kreuz angezeigt werden}

\item{abk}{(logisch) sollen Spaltennamen gekürzt werden}

\item{diagonale}{(logisch) soll die Diagonale gefüllt werden}

\item{...}{andere Parameter für \link{kable}, falls type="html" oder "markdown"}
}
\value{
text table
}
\description{
die Funktion errechnet eine Korrelationstabelle aus den übermittelten Variablen und fügt, wo angebracht, Signifikanzsterne hinzu
}
\examples{
data(pers_data)
corstarsmd_d(pers_data[c(7:10,12:15)],type="ohne", stellen=2, zehnproz=TRUE, untere=FALSE)

\dontrun{
# um das Ergebnis z.B. in Microsoft Word einzufügen, folgende Schritte ausführen:
for_word <- corstarsmd_d(pers_data,type = "ohne",stellen=2) # Ausgabe in Objekt sichern
library(xlsx) # Excel library laden
write.xlsx(for_word, "test.xlsx") # Ausgabe in ein Excel-Dokument speichern
# Das gespeicherte Dokument kann in Excel geöffnet werden, 
# von dort Tabelle kopieren und in Word einfügen
}
}
\seealso{
\link{cor}
}
\keyword{korrelation}
