% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umittr-package.R
\docType{data}
\name{crime_data}
\alias{crime_data}
\title{Verbrechensdaten der USA 1993}
\format{A data frame with 51 rows and 9 variables}
\source{
A. Agresti and B. Finlay, STATISTICAL METHODS FOR THE SOCIAL SCIENCES, Prentice Hall, 3rd edition, 2008
}
\usage{
data(crime_data)
}
\description{
The data are from Statistical Abstract of the United States and most
variables were measured in 1993.
}
\details{
\itemize{
  \item \emph{sid} ID Nummer
  \item \emph{state} US-Staat Abkürzung
  \item \emph{crime} Gewaltverbrechen-Rate (pro 100000)
  \item \emph{murder} Mord-Rate (pro 100000)
  \item \emph{pctmetro} Anteil der Stadtbevölkerung in Prozent
  \item \emph{pctwhite} Anteil der Weißen in Prozent
  \item \emph{pcths} Anteil der Highschool-Absolventen in Prozent
  \item \emph{poverty}  Anteil der als arm einzustufenden Bevölkerung in Prozent
  \item \emph{single} Anteil der Alleinerziehenden in Prozent
}
}
\keyword{datasets}
