% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/extractlavaan2table_function.R
\name{extractlavaan2table}
\alias{extractlavaan2table}
\title{given a lavaan multigroup object returns a dataframe to be used in publication}
\usage{
extractlavaan2table(inva = result)
}
\arguments{
\item{inva}{a lavaan object of an invariance test}
}
\value{
dataframe
}
\description{
<full description>
}
\examples{
require(semTools)
require(lavaan)
data(laurasdaten)
model_reak <- '
reakt =~ RE01_01  + RE01_02  + RE01_03  + RE01_04 + RE01_05 + RE01_06 +
RE01_07 + RE01_08 + RE01_09  + RE01_10 + RE01_11 + RE01_12 + RE01_13 +
RE01_14 + RE01_15 + RE01_16 + RE01_17 + RE01_18
RE01_05 ~~ RE01_10
RE01_05 ~~ RE01_07'
inva <- measurementInvariance(model_reak,laurasdaten,group="Geschlecht",estimator="WLSMV")
extractlavaan2table(inva)
\dontrun{
# if you want e.g. latex output
gsub(" 0.000", " <.001",print(xtable(extractlavaan2table(inva),digits=3), type="latex", sanitize.text.function=identity),useBytes=T)
}
}
\seealso{
lavaan
}
\keyword{invariance}
\keyword{lavaan,}

