% Generated by roxygen2 (4.0.2): do not edit by hand
\name{fa.skalenbilden}
\alias{fa.skalenbilden}
\title{Extrahiert automatisch Skalen aus einem EFA Objekt}
\usage{
fa.skalenbilden(faobject, ladungsunterschied = 0.1, datensatz = NULL,
  name = "skala", fuerntratt = F)
}
\arguments{
\item{faobject}{Objekt aus einer \link{fa} o.ä.}

\item{ladungsunterschied}{Nebenladungsunterschied, der noch akzeptabel ist für das Beibehalten eines Items}

\item{datensatz}{Datensatz (data.frame), der die Items enthält}

\item{name}{Name (Präfix) der Skalen}

\item{fuerntratt}{(logisch) Soll das Fürntratt-Kriterium bei der Faktorzuweiseung angewandt werden}
}
\value{
list of scales
}
\description{
given the pattern maxtrix of an EFA with more than one factor, combine items loading on one particular factor to a scale and save it in an object. Items with unclear loading pattern (i.e. non trivial loadings on more than one factor) can be either evaluated using a loading difference to the next highest loading (parameter loadingdifference) or using the Fürntratt criterion. If an item fails to meet the selected criterion it will not be included in the scale.
}
\examples{
\dontrun{

}
}
\seealso{
\link{fa}, \link{factanal}
}
\keyword{fa}

