% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/row.has.na.R
\name{fall.hat.missings}
\alias{fall.hat.missings}
\title{Missinganalyse pro Zeile}
\usage{
fall.hat.missings(data, welcher = TRUE, zeile = FALSE)
}
\arguments{
\item{data}{Datensatz}

\item{which}{logisch: sollen nur die Fallnummern ausgegeben werden}

\item{Zeile}{logisch: Missings pro Zeile anzeigen}
}
\value{
Fallnummern bzw. Ja/Nein
}
\description{
Missinganalyse pro Zeile
}
\examples{
demoframe <- data.frame(sapply(1:15,function(x) rnorm(50,100,10)))
demoframe <- as.data.frame(lapply(demoframe, function(x) "is.na<-"(x, sample(seq(x), floor(length(x) * runif(1, 0, .2))))))
fall.hat.missings(demoframe)
}

