% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/umittr-package.R
\docType{data}
\name{ist_data}
\alias{ist_data}
\title{Intelligence, School grades and life data}
\format{A data frame with 128 rows and 39 variables}
\usage{
data(ist)
}
\description{
The data comprises items from the IST-2000-R
its verbal, figural, numerical intelligence scores
school grades from math, physics, german, etc. and GPA
plus life data like age, sex, location
}
\details{
\itemize{
  \item vp \emph{VP ID}
  \item sex \emph{Geschlecht} (FACTOR)
  \item age \emph{Alter}
  \item verbint \emph{verbale Intelligenz}
  \item numint \emph{num Intelligenz}
  \item figint \emph{fig Intelligenz}
  \item schlusint \emph{reasoning, schlussfolgerndes Denkes}
  \item se  an	gem	zr	re	rz	fa	wue	ma \emph{9 tests from the IST}
  Satzergaenzung, Analogien, Gemeinsamkeiten, Zahlenreihen, Rechenaufgaben, Rechenzeichen, Figurenauswahl, Wuerfelaufgaben, Matrizen
  \item geb_jahr Geburtsjahr
  \item abi Abischnitt
  \item bundesla	Bundesland (FACTOR)
  \item schulj	Anzahl der Schuljahre
  \item deu_m	Deutsch Note
  \item engl_m	Englisch Note
  \item math_m	Mathe Note
  \item phys_m	Physik Note
  \item chem_m	Chemie Note
  \item bio_m Bio Note
  \item 4 items of working memory (STOPRO, COORD, SWITCH, ATTENT)
  \item verbintsw \emph{verbale Intelligenz Standardwert}
  \item numintsw \emph{numerische Intelligenz Standardwert}
  \item figintsw \emph{figurale Intelligenz Standardwert}
  \item schluintsw \emph{schlussfolgerndes Denkes Standardwert}
  \item z_Auf  z_Koord z-transformierte Aufmerksamkeit und Koordination
  \item Produkt Produktterm z_Auf*z_Koord
  \item Produkt2 Produktterm ATTENT*COORD
  }
}
\keyword{datasets}

