% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/KI.R
\name{KI}
\alias{KI}
\title{Berechnung von Konfidenzintervallen}
\usage{
KI(wert, standardabweichung, reliabilitaet, seitigkeit = 1,
  sicherheit = 0.8, mittelwert = NA, typ = "a")
}
\arguments{
\item{wert}{Roh- bzw. Standardwert im Test}

\item{standardabweichung}{Die Standardabweichung des verwendeten Werts}

\item{reliabilitaet}{Reliabilität des Tests - zwischen 0 und 1}

\item{seitigkeit}{1 oder 2}

\item{sicherheit}{Sicherheitswahrscheinlichkeit zwischen 0 und 1}

\item{mittelwert}{für Äquivalenzhypothese: Mittelwert der Verteilung}

\item{typ}{"a" Äquivalenzhypothese "r" Regressionshypothese}
}
\value{
Konfidenzintervall
}
\description{
Berechnung von Konfidenzintervallen
}
\examples{
KI(wert = 66,standardabweichung = 10,reliabilitaet = .98,sicherheit = .99,seitigkeit = 1,mittelwert = 50,typ = "r") # inklusive Regressionshypothese
}

