% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_output_spss_d.R, R/lm_output_spss_d_n.R
\name{lm_coef_spss_d}
\alias{lm_coef_spss_d}
\title{Koeffizienten einer Lineare Regression dargestellt aehnlich SPSS}
\usage{
lm_coef_spss_d(fit, sterne = TRUE, stellen = 2, fix = TRUE)

lm_coef_spss_d(fit, sterne = TRUE, stellen = 2, fix = TRUE)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}

\item{sterne}{(logisch) sollen Signifikanzsterne (\code{\link{symnum}}) gedruckt werden}

\item{stellen}{(nummerisch) auf wie viele Stellen soll gerundet werden, voreingestellt ist 2}

\item{fix}{(logisch) sollen p Werte gerundet 0 als <.001 angezeigt werden}

\item{fit}{Objekt eines linearen Modells}

\item{sterne}{(logisch) sollen Signifikanzsterne (\code{\link{symnum}}) gedruckt werden}

\item{stellen}{(nummerisch) auf wie viele Stellen soll gerundet werden, voreingestellt ist 2}

\item{fix}{(logisch) sollen p Werte gerundet 0 als <.001 angezeigt werden}
}
\value{
dataframe des Ergebnisses einer Regression ähnlich der in SPSS

dataframe des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
Diese Funktion liefert die Tabelle der Regressionskoeffizienten (\code{β}) ähnlich der Darstellung in SPSS

Diese Funktion liefert die Tabelle der Regressionskoeffizienten (\code{β}) ähnlich der Darstellung in SPSS
}
\examples{
data(crime_data)
fit <- lm(crime ~ pctwhite + pcths + pctmetro + single + poverty,data=crime_data)
lm_coef_spss_d(fit)
\dontrun{
lapply(ll, lm_output_spss) # get outputs from several lm objects
}
data(crime_data)
fit <- lm(crime ~ pctwhite + pcths + pctmetro + single + poverty,data=crime_data)
lm_coef_spss_d(fit)
\dontrun{
lapply(ll, lm_output_spss) # get outputs from several lm objects
}
}
\seealso{
\code{\link{lm}} \code{\link{lm_output_spss_wrapper_d}}

\code{\link{lm}} \code{\link{lm_output_spss_wrapper_d}}
}
\keyword{lm}
\keyword{regression}
\keyword{spss}
