% Generated by roxygen2 (4.0.1): do not edit by hand
\name{lm_coef_spss_d}
\alias{lm_coef_spss_d}
\title{Koeffizienten einer Lineare Regression dargestellt aehnlich SPSS}
\usage{
lm_coef_spss_d(fit)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}
}
\value{
dataframe des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
Diese Funktion liefert die Tabelle der Regressionskoeffizienten (β) ähnlich der Darstellung in SPSS
}
\examples{
\dontrun{

}
}
\seealso{
\code{\link{lm}}
}
\keyword{lm,}
\keyword{regression,}
\keyword{spss}

