% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm_output_spss_d.R
\name{lm_coef_spss_d}
\alias{lm_coef_spss_d}
\title{Koeffizienten einer Lineare Regression dargestellt aehnlich SPSS}
\usage{
lm_coef_spss_d(fit, sterne = TRUE, stellen = 2)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}

\item{sterne}{(logisch) sollen Signifikanzsterne (\code{\link{symnum}}) gedruckt werden}
}
\value{
dataframe des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
Diese Funktion liefert die Tabelle der Regressionskoeffizienten (beta) ähnlich der Darstellung in SPSS
}
\examples{
data(crime_data)
fit <- lm(crime ~ pctwhite + pcths + pctmetro + single + poverty,data=crime_data)
lm_coef_spss_d(fit)
\dontrun{
lapply(ll, lm_output_spss) # get outputs from several lm objects
}
}
\seealso{
\code{\link{lm}}
}
\keyword{lm}
\keyword{regression}
\keyword{spss}

