% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/lm_colld.R
\name{lm_coll}
\alias{lm_coll}
\title{multi collinearity diagnosis inside a linear regression}
\usage{
lm_coll(l, ...)
}
\arguments{
\item{l}{lm object}

\item{...}{}
}
\value{
data.frame colldiag
}
\description{
multi collinearity diagnosis
}
\examples{
\dontrun{

}
}
\keyword{internal}

