% Generated by roxygen2 (4.0.0): do not edit by hand
\name{lm_output_spss}
\alias{lm_output_spss}
\title{linear regression similar to SPSS}
\usage{
lm_output_spss(fit)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}
}
\value{
dataframe with lm output similar to SPSS
}
\description{
<full description>
}
\examples{
library(car)
l <- lm(mpg~disp+hp+wt+drat, data=mtcars)
lm_output_spss(l)
\dontrun{
lapply(ll, lm_output_spss) # get outputs from several lm objects
}
}
\keyword{lm,}
\keyword{regression,}
\keyword{spss}

