% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lm_output_spss_wrapper.R
\name{lm_output_spss_wrapper}
\alias{lm_output_spss_wrapper}
\title{Linear Regression Summary similar to SPSS}
\usage{
lm_output_spss_wrapper(fit, Rsquare = TRUE, coefficients = TRUE,
  collinearity = TRUE, plot = TRUE)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}

\item{plot}{should histogram of residuals, qq plot and predicted/residual value plot be shown}
}
\value{
dataframe des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
<full description>
}
\examples{
library(car)
fit <- lm(mpg ~ disp + hp + wt + drat, data=mtcars)
lm_output_spss_wrapper(fit)
}
\seealso{
\code{\link{lm}}
}
\keyword{lm}
\keyword{regression}
\keyword{spss}

