% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lm_output_spss_wrapper_d.R
\name{lm_output_spss_wrapper_d}
\alias{lm_output_spss_wrapper_d}
\title{Ergebnis einer linearen Regression aehnlich SPSS}
\usage{
lm_output_spss_wrapper_d(fit, Rsquare = TRUE, coefficients = TRUE,
  collinearity = TRUE, plot = TRUE, runden = 3, sterne = TRUE,
  fix = TRUE)
}
\arguments{
\item{fit}{Objekt eines linearen Modells}

\item{Rsquare}{(logisch) soll das R^2 berechnet werden}

\item{coefficients}{(logisch) sollen die Koeffizienten angezeigt werden}

\item{collinearity}{(logisch) soll eine Kollinearitätsdiagnose durchgeführt werden}

\item{plot}{(logisch) sollen Histogramm der Residuen, qq plot und Plot der vorhergesagten Werte/Fehler erstellt werden}

\item{runden}{(nummerisch) Anzahl der Stellen, auf die die Ergebnisse gerundet werden. Sinnvoll sind 2 oder 3}

\item{sterne}{(logisch) sollen Signifikanzsterne angegeben werden}

\item{fix}{(logisch) sollen p Werte gerundet 0 als <.001 angezeigt werden}
}
\value{
Liste mit 3 Elementen (dataframe) des Ergebnisses einer Regression ähnlich der in SPSS
}
\description{
Dreigeteiltes Ergebnis einer linearen Regression ähnlich der Ausgabe in SPSS.  Neben einem Modelübersicht mit Varianzaufklärung werden die Koeffizienten aufgeführt und eine Kollinearitätsdiagnose angeboten.
}
\examples{
library(car)
fit <- lm(mpg ~ disp + hp + wt + drat, data=mtcars) 
lm_output_spss_wrapper_d(fit, runden=2)

\dontrun{
# um "schöne" Tabellen zu bekommen
kable(lm_output_spss_wrapper_d(l))
lapply(ll, function(x) kable(lm_output_spss_wrapper_d(x)))
}
}
\seealso{
\code{\link{lm}} \code{\link{lm_coef_spss_d}}
}
\keyword{lm}
\keyword{regression}
\keyword{spss}

