% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/lm.moderation.R
\name{moderation.slopes}
\alias{moderation.slopes}
\title{calculates slopes and intercepts for a moderation and plots a simple graph}
\usage{
moderation.slopes(mod, zsd = 1, mod_name = "Moderator", ...)
}
\arguments{
\item{mod}{the result of a moderation
done like this: lm(y~x*z), so that you have one predictor x
and one moderator z}

\item{zsd}{the standard deviation of z to calculate the slopes}

\item{...}{Parameter for plot (xlab, ylab, main etc.)}
}
\value{
moderation
}
\description{
calculates the slopes and intercepts for a simple moderation analysis from a lm
}
\examples{
data(crime_data)
attach(crime_data)
#crime <- scale(crime)
poverty <- scale(poverty) #poverty-mean(poverty, na.rm=T)
pcths <- scale(pcths) #pcths-mean(pcths, na.rm=T)
zz <- xm*zm
mod <- lm(crime~poverty*pcths)
summary(mod)
moderation.slopes(mod, ylab="crime rate", mod_name = "Highschool pct", xlab="poverty")
detach(crime_data)
}
\seealso{
\code{\link{lm}}
}
\keyword{lm}

