% Generated by roxygen2 (4.0.0): do not edit by hand
\name{multibar}
\alias{multibar}
\title{<brief desc>}
\usage{
multibar(x, variables, labels, xlab = "Antworten", save = TRUE, ...)
}
\arguments{
\item{x}{}

\item{variables}{}

\item{labels}{}

\item{xlab}{}

\item{save}{}
}
\value{
none
}
\description{
<full description>
}
\examples{
\dontrun{
multi
}
}
\seealso{
mulithist
}
\keyword{umit}

