% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multibar}
\alias{multibar}
\title{Multiple separate barplots}
\usage{
multibar(x, variables, labels, xlab = "Antworten", save = TRUE, ...)
}
\arguments{
\item{x}{data.frame}

\item{variables}{vector of variables to be used}

\item{labels}{}

\item{xlab}{}

\item{save}{should pdf be saved to disk}

\item{...}{additional graphic parameters for barplot (e.g., col)}
}
\value{
Generates for each variable in a dataframe a barplot, optionally saved to the wd as a pdf
}
\description{
given a dataframe, produces a barplot for each variable or a subset of variables, which optionally can be saved to disk as a pdf
}
\examples{
data(cars)
multibar(cars)
}
\seealso{
mulithist
}
\keyword{umit}

