% Generated by roxygen2 (4.0.1): do not edit by hand
\name{multihist}
\alias{multihist}
\title{Multiple separate histograms}
\usage{
multihist(x, variables, labels = TRUE, xlab = "Verteilung", save = TRUE,
  ...)
}
\arguments{
\item{x}{data.frame}

\item{variables}{vector of variables to be used}

\item{labels}{character vector of labels}

\item{xlab}{}

\item{save}{should pdf be saved to disk}

\item{...}{additional graphic parameters vor hist (e.g., col)}
}
\value{
none
}
\description{
produces a histogram for each variable in a dataframe or its subset, which optionally can be saved to disk as a pdf
}
\examples{
\dontrun{
multihist(dd)
multibar(dd, 80:82, save = T)
multibar(x2,c(1:3))
multibar(neo_n,5,ylab="häu")
}
}
\seealso{
lillionscage::multibar
}
\keyword{umit}

