% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/multibar_multihist.R
\name{multihist}
\alias{multihist}
\title{Multiple separate histograms}
\usage{
multihist(x, variables, labels = TRUE, xlab = "Verteilung", save = TRUE,
  format = jpeg, ...)
}
\arguments{
\item{x}{data.frame}

\item{variables}{vector of variables to be used, e.g., c(1,4:10)}

\item{labels}{character vector of labels, useful if elaborate labels are stored
in an object separate from the dataframe}

\item{xlab}{}

\item{save}{should pdf be saved to disk}

\item{format}{specify the format in which to save the file (e.g., jpeg, pdf, png, eps)}

\item{...}{additional graphic parameters for hist (e.g., densitiy, col, ylab, etc)}
}
\value{
none
}
\description{
produces a histogram for each variable in a dataframe or its subset, which optionally can be saved to disk as a pdf
}
\examples{
data(crime_data)
multihist(crime_data, save=FALSE, ylab="Freq. #")
\dontrun{
multihist(dd)
multibar(dd, 80:82, save = T)
multibar(x2,c(1:3))
multibar(neo_n,5,ylab="häu")
}
}
\seealso{
\code{\link{multibar}} \code{\link{hist}}
}
\keyword{plot}
\keyword{umit}

