% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pakete_installieren}
\alias{pakete_installieren}
\title{Installieren notwendiger packages}
\usage{
pakete_installieren()
}
\arguments{
\item{keine}{Parameter}
}
\value{
nichts
}
\description{
völlig unnötige Funktion, aber praktisch
}
\examples{
pakete_installieren()
}
\keyword{helper}

