% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/installieren_notwendiger_packages.R
\name{pakete_installieren}
\alias{pakete_installieren}
\title{Installieren notwendiger packages}
\usage{
pakete_installieren()
}
\arguments{
\item{update}{logisch: sollen gleichzeitig alle Pakete aktualisiert werden}
}
\value{
nichts
}
\description{
voellig unnoetige Funktion, aber praktisch
"car", "lmtest", "QuantPsyc", "lavaan", 
"semTools", "semPlot","GPArotation", "HH", "knitr",
 "sjPlot", "pander", "lattice",  "ggvis","ez","dplyr"
früher noch: "psych", "ggplot2", "reshape2", "Hmisc", "perturb",
"Rcmdr"
}
\examples{
pakete_installieren()
}
\keyword{helper}
