% Generated by roxygen2 (4.0.1): do not edit by hand
\name{pakete_installieren}
\alias{pakete_installieren}
\title{Installieren notwendiger packages}
\usage{
pakete_installieren()
}
\arguments{
\item{keine}{Parameter}
}
\value{
nichts
}
\description{
voellig unnoetige Funktion, aber praktisch
"car", "lmtest", "QuantPsyc", "lavaan", "psych", "ggplot2", "reshape2",
"semTools", "semPlot","GPArotation", "HH", "Hmisc", "knitr","MplusAutomation",
"Rcmdr", "sjPlot", "pander", "lattice", "perturb", "ggvis","ez","dplyr"
}
\examples{
pakete_installieren()
}
\keyword{helper}

