% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Normierung.R
\name{RohwertezuNormwerten}
\alias{RohwertezuNormwerten}
\title{Rohwerten Normwerte zuordnen}
\usage{
RohwertezuNormwerten(Rohwerte, Normwerte)
}
\arguments{
\item{Rohwerte}{}

\item{Normwerte}{}
}
\value{
Dataframe
}
\description{
Rohwerten Normwerte zuordnen
}
\examples{
x <- data.frame(x=rnorm(10000)) # Datensatz erzeugen
x$i <- Tscores(x$x) # Normwerte erzeugen
RohwertezuNormwerten(Rohwerte = x["x"],Normwerte = x["i"])
y <- x[RohwertezuNormwerten(Rohwerte = x$x,Normwerte = x["i"])$rows,]
y
}

