% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/setup.R
\name{setup}
\alias{setup}
\title{erstes Setup von R fuer das RProfile}
\usage{
setup()
}
\value{
schreibt eine User-Datei auf die Festplatte. 
Falls die Datei '.Rprofile' schon besteht, wird ein Backup
der alten Daten erzeugt.
}
\description{
Mit dieser einfachen Funktion ohne Parameter wird für den Benutzer in seinem Heimverzeichnis die .Rprofile Datei erstellt, die bei jedem Start von R aufgerufen wird.
Darin sind Voreinstellungen enthalten, die somit nicht mehr jedes Mal per Hand vorgenommen werden müssen.
\itemize{
\item library psych, ggplot2, foreign, dplyr, umittr werden geladen
\item options(stringsAsFactors=FALSE)
\item options(width=75, digits=3) # Zeilenbreite und Kommastellen
\item options(show.signif.stars=TRUE) # Signifikanzsterne
\item o() # öffent im Explorer/Finder das aktuelle Verzeichnis
\item plotstart() # entfernt den Hintergrund für Plots
\item Schriftfamilien für Graphiken
}
}
\examples{
setup()
# wenn es funktioniert hat:
o() # öffnet das aktuelle Verzeichnis
}

