% Generated by roxygen2 (4.0.2): do not edit by hand
\name{simpleboot}
\alias{simpleboot}
\title{simple univariate bootstrap}
\usage{
simpleboot(x, stat, reps = 1000)
}
\arguments{
\item{x}{}

\item{stat}{}

\item{reps}{}
}
\value{
bootstrap
}
\description{
simple bootstrap for simple univariate statistics
}
\examples{
mydata<-rchisq(25,df=3)
simpleboot(mydata,"mean",reps=10000)
}
\seealso{
boot
}
\keyword{bootstrap}

