% Generated by roxygen2 (4.0.1): do not edit by hand
\name{simpleboot_d}
\alias{simpleboot_d}
\title{Einfacher univariater Bootstrap}
\usage{
simpleboot_d(x, stat, reps = 1000)
}
\arguments{
\item{x}{}

\item{stat}{}

\item{reps}{}
}
\value{
bootstrap
}
\description{
einfacher bootstrap für  univariate Statistik
}
\examples{
{
mydata<-rchisq(25,df=3)
simpleboot_d(mydata, "mean", reps=10000)
simpleboot_d(mydata, sd, reps=5000)
}
}
\seealso{
boot
}
\keyword{bootstrap}

