% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Normierung.R
\name{stanine}
\alias{stanine}
\title{Stanine Werte berechnen}
\usage{
stanine(x)
}
\arguments{
\item{x}{}
}
\value{
Dataframe mit den Grenzwerten für die Stanine
}
\description{
Bestimmt die Staninewert-Grenzen eines gegebenen Vektors. Praktisch für die Normierung eines neuen Tests.
Die Normierung erfolgt nach den Vorschlägen von Telt und Stelzl.
}
\examples{
IQ <- rnorm(20000,100,15) # Datensatz erzeugen
stanine(IQ)
}
\references{
Tent, L., Stelzl, I. (1993), Pädagogisch-psychologische Diagnostik: Theoretische und methodische Grundlagen. Göttingen: Hogrefe.
}
\keyword{Normierung}
\keyword{Stanine,}

