% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/Normierung.R
\name{stanine}
\alias{stanine}
\title{Stanine Werte berechnen}
\usage{
stanine(x)
}
\arguments{
\item{x}{}
}
\value{
Dataframe mit den Grenzwerten für die Stanine
}
\description{
Stanine Werte berechnen
}
\examples{
x <- rnorm(10000) # Datensatz erzeugen
stanine(x)
}

