% Generated by roxygen2 (4.1.0): do not edit by hand
% Please edit documentation in R/umittr-package.R
\docType{package}
\name{umittr-package}
\alias{umittr}
\alias{umittr-package}
\title{Helper functions to ease the transition from SPSS to R}
\description{
Helper functions to ease the transition from SPSS to R
}
\author{
\email{macx@lillion.e4ward.com}
}
\seealso{
\code{\link{...}}
}
\keyword{package}

