% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/umittr-package.R
\docType{package}
\name{umittr-package}
\alias{umittr}
\alias{umittr-package}
\title{Hilfsfunktionen f\code{ü}r den \code{Ü}bergang von SPSS zu R}
\description{
Hilfsfunktionen in R, um den \code{Ü}bergang von GUI zu Syntax zu erleichtern (z.B. SPSS zu R)
}
\author{
\email{macx@lillion.e4ward.com}
}
\seealso{
\code{\link{...}}
}
\keyword{package}

